refactor: split 5 large components into smaller focused modules
